backend/src/requirements.txt
djangorestframework
djangorestframework-api-key
markdown
django-filter  
python-dotenv